moved the same code into a common function
